Backend/requirements.txt
torch
ultralytics
fastapi
uvicorn
pymongo
pillow
python-dotenv
passlib
pyjwt
celery
redis
